requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
python-dotenv>=1.0.0
pytest>=7.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
ta>=0.10.0
streamlit>=1.28.0
altair>=5.0.0